Finished putting the header and footer into the index file
