+++ remove-template-extension.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 
-const viewsDirectory = './dist/views';
+// const viewsDirectory = './dist/views';
 const layoutDirectory = './dist/layout';
 const rootDirectory = './dist/';
 
@@ -32,6 +32,6 @@ function renameFile(filePath) {
     }
 }
 
-processDirectory(viewsDirectory);
+// processDirectory(viewsDirectory);
 processDirectory(layoutDirectory);
 processDirectory(rootDirectory);
