+++ src/styles/global.scss
@@ -31,6 +31,7 @@ body {
 main {
   width: 100vw;
   background-color: white;
+  min-height: 70vh;
 }
 
 section {
