+++ write-components-to-html.js
@@ -5,8 +5,7 @@ const viewsDirectory = './src/views/';
 const layoutDirectory = './src/layout/';
 const componentsDirectory = './src/components/';
 const navFile = './src/layout/nav/nav.html';
-const headerFile = './dist/layout/header/header.html';
-const footerFile = './dist/layout/footer/footer.html';
+
 
 function processMultipleDirectories(...directories) {
     directories.forEach((directory) => {
@@ -71,27 +70,29 @@ processMultipleDirectories(viewsDirectory, layoutDirectory);
 
 
 // Write the updated header into index.html
-setTimeout(() => {
-    const indexPath = path.join(__dirname, 'src/index.html');
-    const indexContent = fs.readFileSync(indexPath, 'utf8');
-    const headerContent = fs.readFileSync(headerFile, 'utf8');
-    const footerContent = fs.readFileSync(footerFile, 'utf8');
-    const indexHeaderRegex = /<header class="header">([\s\S]*?)<\/header>/;
-    const indexFooterRegex = /<footer class="footer">([\s\S]*?)<\/footer>/;
-    const indexHeaderMatch = indexHeaderRegex.exec(indexContent);
-    const indexFooterMatch = indexFooterRegex.exec(indexContent);
-    let updatedIndexContent = indexContent;
-    if (indexHeaderMatch) {
-        const indexHeaderReplacement = `<header>${headerContent}</header>`;
-        updatedIndexContent = updatedIndexContent.replace(indexHeaderMatch[0], indexHeaderReplacement);
-    }
-    if (indexFooterMatch) {
-        const indexFooterReplacement = `<footer>${footerContent}</footer>`;
-        updatedIndexContent = updatedIndexContent.replace(indexFooterMatch[0], indexFooterReplacement);
-    }
-    console.log('indexPath: ', indexPath)
-    console.log('updatedIndexContent: ', updatedIndexContent)
-    fs.writeFileSync(indexPath + ".template", updatedIndexContent);
-}, 2000);
+// setTimeout(() => {
+//     const headerFile = './dist/layout/header/header.html';
+//     const footerFile = './dist/layout/footer/footer.html';
+//     const indexPath = path.join(__dirname, 'src/index.html');
+//     const indexContent = fs.readFileSync(indexPath, 'utf8');
+//     const headerContent = fs.readFileSync(headerFile, 'utf8');
+//     const footerContent = fs.readFileSync(footerFile, 'utf8');
+//     const indexHeaderRegex = /<header class="header">([\s\S]*?)<\/header>/;
+//     const indexFooterRegex = /<footer class="footer">([\s\S]*?)<\/footer>/;
+//     const indexHeaderMatch = indexHeaderRegex.exec(indexContent);
+//     const indexFooterMatch = indexFooterRegex.exec(indexContent);
+//     let updatedIndexContent = indexContent;
+//     if (indexHeaderMatch) {
+//         const indexHeaderReplacement = `<header>${headerContent}</header>`;
+//         updatedIndexContent = updatedIndexContent.replace(indexHeaderMatch[0], indexHeaderReplacement);
+//     }
+//     if (indexFooterMatch) {
+//         const indexFooterReplacement = `<footer>${footerContent}</footer>`;
+//         updatedIndexContent = updatedIndexContent.replace(indexFooterMatch[0], indexFooterReplacement);
+//     }
+//     console.log('indexPath: ', indexPath)
+//     console.log('updatedIndexContent: ', updatedIndexContent)
+//     fs.writeFileSync(indexPath + ".template", updatedIndexContent);
+// }, 2000);
